Discriminate header class in resolution from sources in AbstractLazyMemberScope
Similarly to getFirstClassifierDiscriminateHeaders, we select the first
non-header class if possible, otherwise we select just the first class.
This makes sure that a reference will never be resolved to a header
class if the corresponding impl class is present.

KT-15521 is likely not fixed yet however, because there are other scopes
where header classes are not yet discriminated (LazyImportScope,
DeserializedMemberScope)

 #KT-15521 In Progress